- LinearFilterBasedSMCAlgorithm   - eliminada la parte de muestrear de manera uniforme durante los primeros instantes de tiempo - com.cpp   - eliminados del vector de algoritmos a ejecutar los algoritmos que comparaban distintos thresholds para el algoritmo RLS   - borrado el código para calcular la pe resolviendo la ambiguedad a traves de la matriz de canal final estimada - StatUtil   - añadido método WithoutReplacementSampling (PARECE que funciona) - ViterbiPath   - corregido bug gordísimo en el constructor copia - añadido algoritmo StochasticPSP - eliminados algunos warnings de comparaciones entre signed y unsigned

+++ trunk/smc/resampling/ResidualResamplingAlgorithm.cpp
@@ -38,7 +38,7 @@ std::vector<int> ResidualResamplingAlgorithm::ObtainIndexes(int n,const tVector 
 	int nDeterministicParticles = 0;
 	for(int iWeight=0;iWeight<weights.size();iWeight++)
 	{
-		timesToBeResampled[iWeight] = n*weights(iWeight);
+		timesToBeResampled[iWeight] = int(n*weights(iWeight));
 		nDeterministicParticles += timesToBeResampled[iWeight];
 		residues(iWeight) = double(n)*weights(iWeight) - double(timesToBeResampled[iWeight]);
 	}
